requirements.txt
asgiref==3.9.0
Django==5.2.4
django-model-utils==5.0.0
djangorestframework==3.16.0
drf-secure-token==1.2.0
pillow==11.3.0
psycopg==3.2.9
psycopg2-binary==2.9.10
PyMySQL==1.1.1
sqlparse==0.5.3
typing_extensions==4.14.0
